+++ src/modules/increment_mod.f90
@@ -154,12 +154,7 @@ CONTAINS
     !- Read the analysis mask (in LAM mode only) - N.B. different from land/sea mask!!!
     !
     if (.not. hco_trl%global .and. useAnalIncMask) then
-      call gsv_allocate(statevector_mask, 1, hco_trl, vco_trl, dateStamp_opt=-1, &
-                        dataKind_opt=pre_incrReal, &
-                        mpi_local_opt=.true., varNames_opt=(/'MSKC'/),           &
-                        hInterpolateDegree_opt=hInterpolationDegree)
-      call gsv_readFromFile(statevector_mask, './analinc_mask', ' ', ' ', unitConversion_opt=.false., &
-                            vcoFileIn_opt=vco_trl)
+      call gsv_getMaskLAM(statevector_mask, hco_trl, vco_trl, hInterpolationDegree)
     end if
 
     !
@@ -232,10 +227,10 @@ CONTAINS
     if (.not. hco_trl%global .and. useAnalIncMask) then
       if (gsv_varExist(varName='P0')) then
         call inc_interpolateAndAdd(statevectorIncLowRes, stateVectorAnalHighRes, &
-                                   PsfcReference_opt=PsfcAnalysis(:,:,1,:), mask2d_opt=statevector_mask)
+                                   PsfcReference_opt=PsfcAnalysis(:,:,1,:), statevectorMaskLAM_opt=statevector_mask)
       else
         call inc_interpolateAndAdd(statevectorIncLowRes, stateVectorAnalHighRes, &
-                                   mask2d_opt=statevector_mask)
+                                   statevectorMaskLAM_opt=statevector_mask)
       end if
     else
       if (gsv_varExist(varName='P0')) then
@@ -588,26 +583,25 @@ CONTAINS
   ! inc_interpolateAndAdd
   !--------------------------------------------------------------------------
   subroutine inc_interpolateAndAdd(statevector_in,statevector_inout,scaleFactor_opt, &
-                                   PsfcReference_opt, mask2d_opt)
+                                   PsfcReference_opt, statevectorMaskLAM_opt)
     !
     ! :Purpose: Interpolate the low-resolution increments to trial grid and add to 
     !           get the high-resolution analysis.
     !
     implicit none
-    type(struct_gsv)  :: statevector_in, statevector_inout
 
-    real(8), optional :: scaleFactor_opt
-    real(pre_incrReal), optional :: PsfcReference_opt(:,:,:)
-    type(struct_gsv),   optional :: mask2d_opt
+    ! Arguments:
+    type(struct_gsv), intent(in)    :: statevector_in
+    type(struct_gsv), intent(inout) :: statevector_inout
+    real(8), intent(in), optional   :: scaleFactor_opt
+    real(pre_incrReal), intent(in), optional :: PsfcReference_opt(:,:,:)
+    type(struct_gsv), intent(in), optional   :: statevectorMaskLAM_opt
 
+    ! Locals:
     type(struct_gsv) :: statevector_in_hvInterp
-
     real(pre_incrReal), pointer :: increment(:,:,:,:)
-    real(pre_incrReal), pointer :: analIncMask(:,:,:)
     real(4), allocatable        :: PsfcReference_r4(:,:,:)
     real(8), allocatable        :: PsfcReference_r8(:,:,:)
-
-    integer :: latIndex,kIndex,lonIndex, stepIndex
 
     character(len=4), pointer :: varNamesToInterpolate(:)
 
@@ -660,22 +654,8 @@ CONTAINS
     !
     !- Masking
     !
-    if (present(mask2d_opt)) then
-      call gsv_getField(statevector_in_hvInterp,increment)
-      call gsv_getField(mask2d_opt,analIncMask)
-      do stepIndex = 1, statevector_in_hvInterp%numStep
-        !$OMP PARALLEL DO PRIVATE (latIndex,kIndex,lonIndex)
-        do kIndex = 1, statevector_in_hvInterp%nk
-          do latIndex =  statevector_in_hvInterp%myLatBeg,  statevector_in_hvInterp%myLatEnd
-            do lonIndex =  statevector_in_hvInterp%myLonBeg,  statevector_in_hvInterp%myLonEnd
-              increment(lonIndex,latIndex,kIndex,stepIndex) =      &
-                   increment(lonIndex,latIndex,kIndex,stepIndex) * &
-                   analIncMask(lonIndex,latIndex,1)
-            end do
-          end do
-        end do
-        !$OMP END PARALLEL DO
-      end do
+    if (present(statevectorMaskLAM_opt)) then
+      call gsv_applyMaskLAM(statevector_in_hvInterp,statevectorMaskLAM_opt)
     end if
 
     !